Save information about a directory or data file.

If you give it a directory, it saves a file in the directory. If you give it a filename, it saves a file next to that one.

save_source(path, properties)

Arguments

path
The path to the directory or file.

properties
A list of information about that path.

Details

The format is TOML. The file ends in .rampmd (?). The document structure follows W3C-Prov ontology, as best we can, which means the TOML has subheadings for things that are agents, activities, and entities.

I would like to fill out some properties automatically, when possible.

Some well-known properties:

title Something you would call this data.

download_date When the file was

creator Person or organization that made the file.

creator_email Email of person or organization that made the file.

obtainer Person who got the file.

obtainer_email Email of person who got the file.

format Describe the data format.

source_repository The git repository of code that made the data.

source_version Version number of the source code.

source_hash Hash of the git checkout.

source_branch Branch of the code that made the data.

description A free text description of what's in the file.

creation_date When the file was created.